chore: update snap
