+++ test/rules/no-array-push.project-service.test.ts
@@ -1,0 +1,17 @@
+import path from 'node:path'
+import tsParser from '@typescript-eslint/parser'
+import { run } from '../../src'
+import { common } from './no-array-push.cases'
+
+// Using `projectService` - default project.
+run({
+  ...common,
+  languageOptions: {
+    parser: tsParser,
+    parserOptions: {
+      projectService: {
+        allowDefaultProject: ['*.ts'],
+      },
+    },
+  },
+})
